Allow two-phase borrows of &mut self in ops
We need two-phase borrows of ops to be in the initial NLL release since without
them lots of existing code will break. Fixes #48129
